rename Bar and Foo
