--- extensions/basic/jruby-ext/src/main/java/com/purbon/Foo.java
@@ -1,28 +1,0 @@
-package com.purbon;
-
-import org.jruby.Ruby;
-import org.jruby.RubyClass;
-import org.jruby.RubyObject;
-import org.jruby.RubyString;
-import org.jruby.anno.JRubyMethod;
-import org.jruby.anno.JRubyModule;
-import org.jruby.runtime.ThreadContext;
-import org.jruby.runtime.builtin.IRubyObject;
-
-/**
- * A basic module named Foo
- * Created by purbon on 24/08/15.
- */
-@JRubyModule( name = "Foo")
-public class Foo extends RubyObject {
-
-    public Foo(Ruby ruby, RubyClass metaclass) {
-        super(ruby, metaclass);
-    }
-
-    @JRubyMethod( module = true, name = { "build_string", "new_string" } )
-    public static RubyString buildAnString(ThreadContext context, IRubyObject self) {
-        Ruby runtime = context.runtime;
-        return runtime.newString("This is a new String");
-    }
-}
